refactor(package): move selector family logic out of searchbar component
