+++ demo-todo/src/styles/styles.css
@@ -208,5 +208,11 @@ button span {
 
 /* ---------- Search ---------- */
 .searchContainer {
-  margin: auto;
+  margin-top: 100px;
+  background-color: rgb(63, 63, 63);
+  box-shadow: 0px 0px 35px 20px rgba(10, 10, 10, 0.096);
+  border-radius: 5px;
+  padding: 5px 12px 0 12px;
+  display: flex;
+  justify-content: space-between;
 }
